refactor: Split CLI module
